add static test cases
