+++ creator-node/test/assets/static_text.txt
@@ -1,0 +1,1 @@
+is mayonnaise an instrument
