--- creator-node/test/ipfsClient.test.js
@@ -1,228 +1,0 @@
-const assert = require('assert')
-const fs = require('fs')
-const path = require('path')
-const uuid = require('uuid')
-const sinon = require('sinon')
-
-const ipfsClient = require('../src/ipfsClient')
-const ipfsAdd = require('../src/ipfsAdd')
-
-const ipfs = ipfsClient.ipfs
-const ipfsLatest = ipfsClient.ipfsLatest
-
-const createRandomText = () => {
-  return uuid.v4()
-}
-
-const createRandomTextFile = () => {
-  const filePath = path.join(__dirname, './assets/random.txt')
-  fs.writeFileSync(filePath, createRandomText())
-
-  return filePath
-}
-
-describe('test ipfsClient', () => {
-  let randomText, randomTextFilePath
-
-  before(() => {
-    randomText = createRandomText()
-    randomTextFilePath = createRandomTextFile()
-  })
-
-  afterEach(() => {
-    sinon.restore()
-  })
-
-  // TEST IPFS ADD
-
-  it('[buffer] when adding a buffer, the cids should be consistent', async () => {
-    const buffer = fs.readFileSync(randomTextFilePath)
-
-    // Original IPFS instance
-    const ipfsAddBufferResp = (await ipfs.add(buffer))[0].hash
-
-    // Original IPFS instance with `onlyHash` flag
-    const ipfsAddBufferRespOnlyHashFlag = (await ipfs.add(buffer, { onlyHash: true }))[0].hash
-
-    // Newer IPFS instance
-    let ipfsLatestAddBufferResp
-    for await (const resp of ipfsLatest.add(buffer)) {
-      ipfsLatestAddBufferResp = `${resp.cid}`
-    }
-
-    // Newer IPFS instance with `onlyHash` flag
-    let ipfsLatestAddBufferRespWithOnlyHashFlag
-    for await (const resp of ipfsLatest.add(buffer, { onlyHash: true })) {
-      ipfsLatestAddBufferRespWithOnlyHashFlag = `${resp.cid}`
-    }
-
-    // Only hash logic
-    const onlyHash = await ipfsAdd.ipfsOnlyHashNonImages(buffer)
-
-    console.log('buffers', {
-      ipfsAddBufferResp,
-      ipfsAddBufferRespOnlyHashFlag,
-      ipfsLatestAddBufferResp,
-      ipfsLatestAddBufferRespWithOnlyHashFlag,
-      onlyHash
-    })
-
-    const allResults = [
-      ipfsAddBufferResp,
-      ipfsAddBufferRespOnlyHashFlag,
-      ipfsLatestAddBufferResp,
-      ipfsLatestAddBufferRespWithOnlyHashFlag,
-      onlyHash
-    ]
-
-    assert.ok(!!allResults.reduce(function (a, b) { return (a === b) ? a : NaN }))
-  })
-
-  it('[readstream] when adding a readstream, the cids should be consistent', async () => {
-    // Original IPFS instance
-    const ipfsAddReadstreamResp = (await ipfs.add(fs.createReadStream(randomTextFilePath)))[0].hash
-
-    // Original IPFS instance with `onlyHash` flag
-    const ipfsAddReadstreamRespOnlyHashFlag = (await ipfs.add(fs.createReadStream(randomTextFilePath), { onlyHash: true }))[0].hash
-
-    // Newer IPFS instance
-    let ipfsLatestAddReadstreamResp
-    for await (const resp of ipfsLatest.add(fs.createReadStream(randomTextFilePath))) {
-      ipfsLatestAddReadstreamResp = `${resp.cid}`
-    }
-
-    // Newer IPFS instance with `onlyHash` flag
-    let ipfsLatestAddReadstreamRespWithOnlyHashFlag
-    for await (const resp of ipfsLatest.add(fs.createReadStream(randomTextFilePath), { onlyHash: true })) {
-      ipfsLatestAddReadstreamRespWithOnlyHashFlag = `${resp.cid}`
-    }
-
-    // Only hash logic
-    const onlyHash = await ipfsAdd.ipfsOnlyHashNonImages(fs.createReadStream(randomTextFilePath))
-
-    console.log('readstream', {
-      ipfsAddReadstreamResp,
-      ipfsAddReadstreamRespOnlyHashFlag,
-      ipfsLatestAddReadstreamResp,
-      ipfsLatestAddReadstreamRespWithOnlyHashFlag,
-      onlyHash
-    })
-
-    // NOTE: we have observed that passing in a readstream to ipfs.add is inconsistent.
-    // Commenting out ipfs.add responses in test result comparisons
-    const allResults = [
-      // ipfsAddReadstreamResp,
-      // ipfsAddReadstreamRespOnlyHashFlag,
-      ipfsLatestAddReadstreamResp,
-      ipfsLatestAddReadstreamRespWithOnlyHashFlag,
-      onlyHash
-    ]
-
-    // Check that every response in `allResults` is equal to each other
-    assert.ok(!!allResults.reduce(function (a, b) { return (a === b) ? a : NaN }))
-  })
-
-  // Note: ipfsLatest does not have a `addFromFs()` fn
-  it('[file path] when adding a file path, the cids should be consistent', async () => {
-    // Original IPFS instance
-    const ipfsAddFromFsFilePathResp = (await ipfs.addFromFs(randomTextFilePath))[0].hash
-
-    // Original IPFS instance with `onlyHash` flag
-    const ipfsAddFromFsFilePathRespOnlyHashFlag = (await ipfs.addFromFs(randomTextFilePath, { onlyHash: true }))[0].hash
-
-    // Only hash logic
-    const buffer = await fs.readFileSync(randomTextFilePath)
-    const onlyHash = await ipfsAdd.ipfsOnlyHashNonImages(buffer)
-
-    console.log('file path', {
-      ipfsAddFromFsFilePathResp,
-      ipfsAddFromFsFilePathRespOnlyHashFlag,
-      onlyHash
-    })
-
-    const allResults = [
-      ipfsAddFromFsFilePathResp,
-      ipfsAddFromFsFilePathRespOnlyHashFlag,
-      onlyHash
-    ]
-
-    assert.ok(!!allResults.reduce(function (a, b) { return (a === b) ? a : NaN }))
-  })
-
-  it('[ipfsAddNonImages] passing in a Buffer and `enableIPFSAdd` = true should work', async () => {
-    const buffer = Buffer.from(randomText)
-
-    const onlyHash = await ipfsAdd.ipfsAddNonImages(buffer)
-    const ipfsLatestAddWithDaemonResp = await ipfsAdd.ipfsAddNonImages(buffer, {}, {}, true)
-
-    console.log('buffer', {
-      onlyHash,
-      ipfsLatestAddWithDaemonResp
-    })
-
-    assert.deepStrictEqual(onlyHash, ipfsLatestAddWithDaemonResp)
-  })
-
-  it('[ipfsAddNonImages] passing in a ReadStream and `enableIPFSAdd` = true should work', async () => {
-    const readStream1 = fs.createReadStream(randomTextFilePath)
-    const onlyHash = await ipfsAdd.ipfsAddNonImages(readStream1)
-
-    const readStream2 = fs.createReadStream(randomTextFilePath)
-    const ipfsLatestAddWithDaemonResp = await ipfsAdd.ipfsAddNonImages(readStream2, {}, {}, true)
-
-    console.log('readstream', {
-      onlyHash,
-      ipfsLatestAddWithDaemonResp
-    })
-
-    assert.deepStrictEqual(onlyHash, ipfsLatestAddWithDaemonResp)
-  })
-
-  it('[ipfsAddNonImages] passing in a source path and `enableIPFSAdd` = true should work', async () => {
-    const pathToFile = path.join(__dirname, './assets/random.txt')
-
-    const onlyHash = await ipfsAdd.ipfsAddNonImages(pathToFile)
-    const ipfsLatestAddWithDaemonResp = await ipfsAdd.ipfsAddNonImages(pathToFile, {}, {}, true)
-
-    console.log('source path', {
-      onlyHash,
-      ipfsLatestAddWithDaemonResp
-    })
-
-    assert.deepStrictEqual(onlyHash, ipfsLatestAddWithDaemonResp)
-  })
-
-  it('[ipfsAddNonImages] passing in improper content throws', async () => {
-    try {
-      await ipfsAdd.ipfsAddNonImages('vicky was here')
-      throw new Error('passing in improper data should have failed')
-    } catch (e) {
-      assert.ok(e.toString().includes('Could not convert content into buffer'))
-    }
-  })
-
-  it('[ipfsAddNonImages] if `ipfsOnlyHashImages()` errors, `ipfsAddNonImages()` throws', async () => {
-    const pathToFile = path.join(__dirname, './assets/random.txt')
-    sinon.stub(ipfsAdd, 'ipfsOnlyHashNonImages').throws(new Error('failed to generate only hash'))
-
-    try {
-      await ipfsAdd.ipfsAddNonImages(pathToFile)
-      throw new Error('`ipfsAddNonImages` should throw if `ipfsOnlyHashNonImages` fails')
-    } catch (e) {
-      assert.ok(e.toString().includes('failed to generate only hash'))
-    }
-  })
-
-  it('[ipfsAddNonImages] if `ipfsOnlyHashImages()` and ipfs daemon hash resp diverges, `ipfsAddNonImages()` throws', async () => {
-    const pathToFile = path.join(__dirname, './assets/random.txt')
-    sinon.stub(ipfsAdd, 'ipfsOnlyHashNonImages').returns('QmVHxRocoWgUChLEvfEyDuuD6qJ4PhdDL2dTLcpUy3dSC2'
-    )
-
-    try {
-      await ipfsAdd.ipfsAddNonImages(pathToFile, {}, {}, true)
-      throw new Error('`ipfsAddNonImages` should throw if only hash and daemon hash diverges')
-    } catch (e) {
-      assert.ok(e.toString().includes('are not consistent'))
-    }
-  })
-})
